Remove (subtle) coupling between the layers
`schemaversion` property is specific to the contents of the fetched
json file. And fetcher, which doesn't know where the number is being
obtained from, should not know that property's name.

In order to remove the coupling, the number is being set
using a "contracted" name.

Any other fetcher, that would be used, is now expected
to return an object with `schemaVersionNumber` property
holding that value.