Lots of changes
Generally speaking the whole DataType was rewritten.
DataType helper methods were added. They are not schema-like
(not existing as properties), but it looked like a good place
to store them there.

As this is still an alfa / beta version of this project, I think
I can go with it and see how this will feel like.

I am thinking about extracting those methods to a helper, so that
the DataType schema really reflects what is has (in schema.org).